agent/chat/pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "chat-agent"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "bedrock-agentcore >= 1.0.3",
    "httpx >= 0.27.0",
    "mcp >= 1.19.0",
    "mcp-proxy-for-aws >= 1.1.5",
    "pyyaml >= 6.0",
    "strands-agents >= 1.13.0",
    "strands-agents-tools >= 0.1.0",
]

[project.optional-dependencies]
dev = [
    "chromadb >= 0.4.0",
]